interview_crew/pyproject.toml
[project]
name = "interview_crew"
version = "0.1.0"
description = "interview-crew using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=1.0.0,<2.0.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
interview_crew = "interview_crew.main:run"
run_crew = "interview_crew.main:run"
train = "interview_crew.main:train"
replay = "interview_crew.main:replay"
test = "interview_crew.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
